+++ GieAndVince/Models/ViewModel/OrderCartRemoveViewModel.cs
@@ -9,7 +9,7 @@ namespace GieAndVince.Models.ViewModel
     public class OrderCartRemoveViewModel
     {
         public string Message { get; set; }
-        public decimal OrderTotal { get; set; }
+        public double OrderTotal { get; set; }
         public int OrderCount { get; set; }
         public int ItemCount { get; set; }
         public int DeleteId { get; set; }
